Split basic_product into base_product/scheduled_trip_product
