+++ spec/scheduled_trip_product_spec.rb
@@ -1,7 +1,8 @@
 require 'spec_helper'
-require 'quick_travel/basic_product'
+require 'quick_travel/base_product'
+require 'quick_travel/scheduled_trip_product'
 
-describe QuickTravel::BasicProduct do
+describe QuickTravel::ScheduledTripProduct do
   let(:params) {
     {
       product_type_id: 1,
@@ -15,7 +16,7 @@ describe QuickTravel::BasicProduct do
   let(:date) { '2016-03-01' }
   subject(:products) {
     VCR.use_cassette 'basic_product_scheduled_trips' do
-      QuickTravel::BasicProduct.find(:scheduled_trips, params)
+      QuickTravel::ScheduledTripProduct.find(params)
     end
   }
 
@@ -33,7 +34,7 @@ describe QuickTravel::BasicProduct do
   context 'when date has no services' do
     subject(:unbookable_products) {
       VCR.use_cassette 'basic_product_scheduled_trips_unbookable' do
-        QuickTravel::BasicProduct.find(:scheduled_trips, params)
+        QuickTravel::ScheduledTripProduct.find(params)
       end
     }
     let(:date) { '2016-03-03' }
@@ -57,7 +58,7 @@ describe QuickTravel::BasicProduct do
     let(:date) { '2016-03-01' }
     subject(:products) {
       VCR.use_cassette 'basic_product_scheduled_trips_multi_sector' do
-        QuickTravel::BasicProduct.find(:scheduled_trips, params)
+        QuickTravel::ScheduledTripProduct.find(params)
       end
     }
